frontend/requirements.txt
fastapi==0.63.0
google-auth==1.27.1
gunicorn==20.0.4
requests==2.25.1
uvicorn==0.13.4
